Improve support for virtual env. (#26)
When VIRTUAL_ENV variable is present in the environment, use "python
mypy" instead of "mypy" so that library loaded in the VIRTUAL_ENV get
correctly loaded. This works because python is always present is the
VIRTUAL_ENV, but requires that mypy to be present is the directory
where the caller is. In the context of the python-lsp-server and pipx
install/inject this is the case.