Fix matching column numbers when they are present
In projects which configure 'show_column_numbers = true', pylsp would
not return anything back to my editor.

In the logs I found messages like below

```
... discarding result for beetsplug/bandcamp/track.py:113 against /home/sarunas/repo/beetcamp/beetsplug/bandcamp/track.py
```

The logic only returns the result if the latter filename ends with the
former, thus it seems that the column number (':113') is not supposed to
be present here.

Having had a look at the matching pattern I found that it does not take
into account that the column number may be present.

This commit adjusts the pattern to handle the column number correctly.